% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_internal_migration_summary.R
\docType{data}
\name{nz_internal_migration_summary}
\alias{nz_internal_migration_summary}
\title{New Zealand Internal Migration Summary Statistics 2013}
\format{A data frame with 351972 rows and 11 variables.  This single table
includes data from all the separate tables in the original spreadsheet, so
some of the column names are generic (\code{var1} ... \code{var6}).
\itemize{
\item \code{table} Name of the table from the original spreadsheet
\item \code{title} Title of the table from the original spreadsheet
\item \code{subtitle} Subtitle of the table from the original spreadsheet
\item \code{var1} First variable of the table from the original spreadsheet
\item \code{...}
\item \code{var6} Sixth variable of the table from the original spreadsheet
\item \code{count} Number of people
\item \code{flag} Only one flag is used, \code{..C}, to mean 'confidential'.
}}
\source{
http://m.stats.govt.nz/~/media/Statistics/browse-categories/population/migration/internal-migration-tables/int-mig-2013-summary-tables.xlsx
http://m.stats.govt.nz/browse_for_stats/population/Migration/internal-migration/tables.aspx
}
\usage{
nz_internal_migration_summary
}
\description{
A dataset containing tables of summary statistics describing internal
migration in New Zealand, from the 2013 census.
}
\keyword{datasets}
